[Fix/#92] Category 수정 시 데이터 업데이트 문제 해결
CategoryDetailVM과 CategoryListVM을 CategoryVM으로 통합
